refactor: delete topic_callback
